+++ examples/simple-example/App.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { SnackbarProvider } from 'notistack';
 import MessageButtons from './MessageButtons';
 
-
 export default () => (
     <SnackbarProvider maxSnack={3}>
         <MessageButtons />